backend/requirements.txt
fastapi
uvicorn
datasets
faiss-cpu
sentence-transformers
numpy
openai